[api] Audit Log
The goal is to be able to trace who have done what request in SDA.

Since (almost) all requests are handleded first by the RBAC middleware
It makes sense to put the logic there.

Could be that I have missed somewhere we need these auditlogs not
covered by the rbac middleware.